+++ src/common/utils/profileUtils.ts
@@ -33,14 +33,18 @@ export async function minimizeProfile(profileFile: string) {
       removed.push(node);
     }
   }
-  // Keep only default values
+  // Keep only default values or false values
   let updatedDefaults = false;
   const nodesHavingDefault = ["applicationVisibilities", "recordTypeVisibilities"];
   for (const node of nodesHavingDefault) {
     if (profileXml.Profile[node]) {
       const prevLen = profileXml.Profile[node].length;
       profileXml.Profile[node] = profileXml.Profile[node].filter((nodeVal) => {
-        if ((nodeVal?.default && nodeVal?.default[0] === "true") || (nodeVal?.personAccountDefault && nodeVal?.personAccountDefault[0] === "true")) {
+        if (
+          (nodeVal?.default && nodeVal?.default[0] === "true") ||
+          (nodeVal?.personAccountDefault && nodeVal?.personAccountDefault[0] === "true") ||
+          (nodeVal?.visible && nodeVal?.visible[0] === "false")
+        ) {
           return true;
         }
         return false;